Remove Python notebook text and ImageCleaner section
I got tired of the Pets text that we never removed from way earlier in
the project. Additionally, we are not going to use ImageCleaner
because it is buggy, there is no way to access the GUI interface, and
all of our photos should be valid anyway because we literally
downloaded from a zipfile instead of scraping from the webs.