Increase sample size to 300 and unfreeze layers
The script now runs smoothly on a local machine. It runs the LR finder
if that section is uncommented, then unfreezes the layers and trains
for 20 epochs. All of the plots are now saved as files in the same
directory, and the terminal output (with the exception of the epoch
training results to show training progress) are saved into an output
text file for easy access.